+++ crates/gamai/src/common_actions/timer_actions.rs
@@ -1,7 +1,7 @@
 use crate::*;
 use bevy_ecs::prelude::*;
 use bevy_time::prelude::*;
-
+use std::time::Duration;
 
 /// An action that updates the action timer for a given node.
 /// The timer will update even if the node is not running.
@@ -17,5 +17,22 @@ pub fn update_action_timer<N: AiNode>(
 		}
 	} else if timers.iter().next().is_some() {
 		eprintln!("Gamai - `ActionTimer` found but `Time` resource is missing");
+	}
+}
+
+
+
+#[action]
+pub fn succeed_in_one_second<N: AiNode>(
+	mut commands: Commands,
+	mut query: Query<(Entity, &Prop<ActionTimer, N>), With<Prop<Running, N>>>,
+) {
+	for (entity, timer) in query.iter_mut() {
+		if timer.last_start.elapsed() >= Duration::from_secs(1) {
+			// println!("last start: {:?}", timer.last_start.elapsed());
+			commands
+				.entity(entity)
+				.insert(Prop::<_, N>::new(ActionResult::Success));
+		}
 	}
 }